requirements.txt
mysql-connector==2.2.9
numpy==1.21.0
opencv-contrib-python==4.5.3.56
opencv-python==4.5.3.56
Pillow==8.3.1
